enable playwright tests again
