+++ e2e/testUtils.ts
@@ -1,545 +1,537 @@
 import path from "path";
-import { BrowserContext, chromium, Page } from "playwright";
-import { ethers } from 'ethers';
+import {BrowserContext, chromium, Page} from "playwright";
+import {ethers} from 'ethers';
 import fs from "fs";
 
 export async function handleMetaMaskNetworkAndConfirm(
-  context: BrowserContext,
-  shouldSwitchNetwork: boolean = true,
-  existingPage?: Page
+    context: BrowserContext,
+    shouldSwitchNetwork: boolean = true,
+    existingPage?: Page
 ): Promise<void> {
-  try {
-    // Get MetaMask page (either existing or wait for new one)
-    let metaMaskPage: Page;
-    const existingMetaMaskPage = context.pages().find(page => page.url().includes('extension'));
-
-    if (existingMetaMaskPage && !existingMetaMaskPage.isClosed()) {
-      metaMaskPage = existingMetaMaskPage;
-      console.log("Using existing MetaMask page");
-    } else {
-      console.log("Waiting for new MetaMask page...");
-      try {
-        await context.waitForEvent('page', { timeout: 10000 });
-        metaMaskPage = context.pages().find(page => page.url().includes('extension')) || context.pages()[1];
-      } catch (error) {
-        console.log("Failed to get new MetaMask page:", error);
-        return; // Exit early if we can't get a MetaMask page
-      }
-    }
-
-    // Check if page is still available
-    if (!metaMaskPage || metaMaskPage.isClosed()) {
-      console.log("MetaMask page is not available or was closed");
-      return; // Exit early if page is not available
-    }
-
-    console.log("Current MetaMask page URL:", metaMaskPage.url());
-
-    // Wait for any content to load
     try {
-      await metaMaskPage.waitForLoadState('domcontentloaded', { timeout: 5000 });
-    } catch (error) {
-      console.log("Failed to wait for page load state:", error);
-      if (metaMaskPage.isClosed()) {
-        console.log("MetaMask page was closed during load");
-        return;
-      }
-    }
+        // Get MetaMask page (either existing or wait for new one)
+        let metaMaskPage: Page;
+        const existingMetaMaskPage = context.pages().find(page => page.url().includes('extension'));
 
-    // Handle different possible states in sequence
-    const maxAttempts = 5;
-    let currentAttempt = 0;
-
-    while (currentAttempt < maxAttempts) {
-      // Check if page is still available before each attempt
-      if (metaMaskPage.isClosed()) {
-        console.log(`MetaMask page was closed during attempt ${currentAttempt + 1}`);
-        // return;
-      }
-
-      try {
-        // Check for network permission dialog
-        const permissionButton = metaMaskPage.locator('[data-testid="page-container-footer-next"]');
-        const isPermissionVisible = await permissionButton.isVisible().catch(() => false);
-        if (isPermissionVisible) {
-          console.log(`Attempt ${currentAttempt + 1}: Network permission dialog found`);
-          await permissionButton.click().catch(e => console.log("Failed to click permission button:", e));
-          // Use a try/catch for the timeout to handle page closure
-          try {
-            if (!metaMaskPage.isClosed()) {
-              await metaMaskPage.waitForTimeout(1000);
+        if (existingMetaMaskPage && !existingMetaMaskPage.isClosed()) {
+            metaMaskPage = existingMetaMaskPage;
+            console.log("Using existing MetaMask page");
+        } else {
+            console.log("Waiting for new MetaMask page...");
+            try {
+                await context.waitForEvent('page', {timeout: 10000});
+                metaMaskPage = context.pages().find(page => page.url().includes('extension')) || context.pages()[1];
+            } catch (error) {
+                console.log("Failed to get new MetaMask page:", error);
+                return; // Exit early if we can't get a MetaMask page
             }
-          } catch (e) {
-            console.log("Page closed during timeout - isPermissionVisible");
-            return;
-          }
-          currentAttempt++;
-          continue;
         }
 
-        // Check for network switch dialog
-        if (shouldSwitchNetwork) {
-          const sepoliaText = metaMaskPage.getByText("Sepolia");
-          const isSepoliaVisible = await sepoliaText.isVisible().catch(() => false);
-          if (isSepoliaVisible) {
-            console.log(`Attempt ${currentAttempt + 1}: Network switch dialog found`);
-            const switchButton = metaMaskPage.locator('[data-testid="page-container-footer-next"]');
-            const isSwitchVisible = await switchButton.isVisible().catch(() => false);
-            if (isSwitchVisible) {
-              await switchButton.click().catch(e => console.log("Failed to click switch button:", e));
-              try {
-                if (!metaMaskPage.isClosed()) {
-                  await metaMaskPage.waitForTimeout(1000);
-                }
-              } catch (e) {
-                console.log("Page closed during timeout - shouldSwitchNetwork");
+        // Check if page is still available
+        if (!metaMaskPage || metaMaskPage.isClosed()) {
+            console.log("MetaMask page is not available or was closed");
+            return; // Exit early if page is not available
+        }
+
+        console.log("Current MetaMask page URL:", metaMaskPage.url());
+
+        // Wait for any content to load
+        try {
+            await metaMaskPage.waitForLoadState('domcontentloaded', {timeout: 5000});
+        } catch (error) {
+            console.log("Failed to wait for page load state:", error);
+            if (metaMaskPage.isClosed()) {
+                console.log("MetaMask page was closed during load");
                 return;
-              }
-              currentAttempt++;
-              continue;
             }
-          }
         }
 
-        // Check for the Transfer request dialog with "Review alert" button
-        // First check for the dialog title
-        const transferRequestTitle = metaMaskPage.getByText('Transfer request');
-        const isTransferRequestVisible = await transferRequestTitle.isVisible().catch(() => false);
+        // Handle different possible states in sequence
+        const maxAttempts = 5;
+        let currentAttempt = 0;
 
-        if (isTransferRequestVisible) {
-          console.log(`Attempt ${currentAttempt + 1}: Transfer request dialog found`);
-
-          // Take a screenshot for debugging
-          // try {
-          //   await metaMaskPage.screenshot({ path: 'metamask-transfer-request.png' }).catch(() => { });
-          //   console.log("Saved screenshot of transfer request dialog");
-          // } catch (e) {
-          //   console.log("Failed to take screenshot");
-          // }
-
-          // Try multiple approaches to find and click the Review alert button
-
-          // Approach 1: Try direct text selector
-          let reviewAlertButton = metaMaskPage.locator('button:has-text("Review alert")');
-          let isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
-
-          // Approach 2: Try CSS selector for primary button
-          if (!isReviewAlertVisible) {
-            reviewAlertButton = metaMaskPage.locator('.btn-primary');
-            isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
-          }
-
-          // Approach 3: Try by role
-          if (!isReviewAlertVisible) {
-            reviewAlertButton = metaMaskPage.getByRole('button', { name: /review alert/i });
-            isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
-          }
-
-          // Approach 4: Try by XPath
-          if (!isReviewAlertVisible) {
-            reviewAlertButton = metaMaskPage.locator('//button[contains(text(), "Review")]');
-            isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
-          }
-
-          if (isReviewAlertVisible) {
-            console.log(`Attempt ${currentAttempt + 1}: Review alert button found`);
-            await reviewAlertButton.click().catch(e => console.log("Failed to click review alert button:", e));
-            console.log("Clicked review alert button");
-          } else {
-            // If no specific button found, try clicking any button that might be a confirm button
-            console.log("No specific review alert button found, trying alternative approaches");
-
-            // Try clicking any button that might be a confirm button
-            const buttons = await metaMaskPage.locator('button').all();
-            console.log(`Found ${buttons.length} buttons in the dialog`);
-
-            // Try to find a button that looks like a confirm button (usually the rightmost one)
-            let buttonClicked = false;
-            for (const button of buttons) {
-              const buttonText = await button.textContent().catch(() => "");
-              const isVisible = await button.isVisible().catch(() => false);
-
-              if (isVisible && (buttonText?.includes("Review") || buttonText?.includes("Confirm") || buttonText?.includes("Accept"))) {
-                console.log(`Found button with text: ${buttonText}`);
-                await button.click().catch(e => console.log(`Failed to click button with text ${buttonText}:`, e));
-                console.log(`Clicked button with text: ${buttonText}`);
-                buttonClicked = true;
-                break;
-              }
+        while (currentAttempt < maxAttempts) {
+            // Check if page is still available before each attempt
+            if (metaMaskPage.isClosed()) {
+                console.log(`MetaMask page was closed during attempt ${currentAttempt + 1}`);
+                // return;
             }
 
-            // If no specific button found, try the last visible button (usually the confirm button)
-            if (!buttonClicked && buttons.length > 0) {
-              for (let i = buttons.length - 1; i >= 0; i--) {
-                const isVisible = await buttons[i].isVisible().catch(() => false);
-                if (isVisible) {
-                  console.log("Clicking the last visible button in the dialog");
-                  await buttons[i].click().catch(e => console.log("Failed to click last button:", e));
-                  buttonClicked = true;
-                  break;
+            try {
+                // Check for network permission dialog
+                const permissionButton = metaMaskPage.locator('[data-testid="page-container-footer-next"]');
+                const isPermissionVisible = await permissionButton.isVisible().catch(() => false);
+                if (isPermissionVisible) {
+                    console.log(`Attempt ${currentAttempt + 1}: Network permission dialog found`);
+                    await permissionButton.click().catch(e => console.log("Failed to click permission button:", e));
+                    // Use a try/catch for the timeout to handle page closure
+                    try {
+                        if (!metaMaskPage.isClosed()) {
+                            await metaMaskPage.waitForTimeout(1000);
+                        }
+                    } catch (e) {
+                        console.log("Page closed during timeout - isPermissionVisible");
+                        return;
+                    }
+                    currentAttempt++;
+                    continue;
                 }
-              }
+
+                // Check for network switch dialog
+                if (shouldSwitchNetwork) {
+                    const sepoliaText = metaMaskPage.getByText("Sepolia");
+                    const isSepoliaVisible = await sepoliaText.isVisible().catch(() => false);
+                    if (isSepoliaVisible) {
+                        console.log(`Attempt ${currentAttempt + 1}: Network switch dialog found`);
+                        const switchButton = metaMaskPage.locator('[data-testid="page-container-footer-next"]');
+                        const isSwitchVisible = await switchButton.isVisible().catch(() => false);
+                        if (isSwitchVisible) {
+                            await switchButton.click().catch(e => console.log("Failed to click switch button:", e));
+                            try {
+                                if (!metaMaskPage.isClosed()) {
+                                    await metaMaskPage.waitForTimeout(1000);
+                                }
+                            } catch (e) {
+                                console.log("Page closed during timeout - shouldSwitchNetwork");
+                                return;
+                            }
+                            currentAttempt++;
+                            continue;
+                        }
+                    }
+                }
+
+                // Check for the Transfer request dialog with "Review alert" button
+                // First check for the dialog title
+                const transferRequestTitle = metaMaskPage.getByText('Transfer request');
+                const isTransferRequestVisible = await transferRequestTitle.isVisible().catch(() => false);
+
+                if (isTransferRequestVisible) {
+                    console.log(`Attempt ${currentAttempt + 1}: Transfer request dialog found`);
+
+                    // Take a screenshot for debugging
+                    // try {
+                    //   await metaMaskPage.screenshot({ path: 'metamask-transfer-request.png' }).catch(() => { });
+                    //   console.log("Saved screenshot of transfer request dialog");
+                    // } catch (e) {
+                    //   console.log("Failed to take screenshot");
+                    // }
+
+                    // Try multiple approaches to find and click the Review alert button
+
+                    // Approach 1: Try direct text selector
+                    let reviewAlertButton = metaMaskPage.locator('button:has-text("Review alert")');
+                    let isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
+
+                    // Approach 2: Try CSS selector for primary button
+                    if (!isReviewAlertVisible) {
+                        reviewAlertButton = metaMaskPage.locator('.btn-primary');
+                        isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
+                    }
+
+                    // Approach 3: Try by role
+                    if (!isReviewAlertVisible) {
+                        reviewAlertButton = metaMaskPage.getByRole('button', {name: /review alert/i});
+                        isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
+                    }
+
+                    // Approach 4: Try by XPath
+                    if (!isReviewAlertVisible) {
+                        reviewAlertButton = metaMaskPage.locator('//button[contains(text(), "Review")]');
+                        isReviewAlertVisible = await reviewAlertButton.isVisible().catch(() => false);
+                    }
+
+                    if (isReviewAlertVisible) {
+                        console.log(`Attempt ${currentAttempt + 1}: Review alert button found`);
+                        await reviewAlertButton.click().catch(e => console.log("Failed to click review alert button:", e));
+                        console.log("Clicked review alert button");
+                    } else {
+                        // If no specific button found, try clicking any button that might be a confirm button
+                        console.log("No specific review alert button found, trying alternative approaches");
+
+                        // Try clicking any button that might be a confirm button
+                        const buttons = await metaMaskPage.locator('button').all();
+                        console.log(`Found ${buttons.length} buttons in the dialog`);
+
+                        // Try to find a button that looks like a confirm button (usually the rightmost one)
+                        let buttonClicked = false;
+                        for (const button of buttons) {
+                            const buttonText = await button.textContent().catch(() => "");
+                            const isVisible = await button.isVisible().catch(() => false);
+
+                            if (isVisible && (buttonText?.includes("Review") || buttonText?.includes("Confirm") || buttonText?.includes("Accept"))) {
+                                console.log(`Found button with text: ${buttonText}`);
+                                await button.click().catch(e => console.log(`Failed to click button with text ${buttonText}:`, e));
+                                console.log(`Clicked button with text: ${buttonText}`);
+                                buttonClicked = true;
+                                break;
+                            }
+                        }
+
+                        // If no specific button found, try the last visible button (usually the confirm button)
+                        if (!buttonClicked && buttons.length > 0) {
+                            for (let i = buttons.length - 1; i >= 0; i--) {
+                                const isVisible = await buttons[i].isVisible().catch(() => false);
+                                if (isVisible) {
+                                    console.log("Clicking the last visible button in the dialog");
+                                    await buttons[i].click().catch(e => console.log("Failed to click last button:", e));
+                                    buttonClicked = true;
+                                    break;
+                                }
+                            }
+                        }
+                    }
+
+                    // Wait for the next dialog to appear
+                    try {
+                        if (!metaMaskPage.isClosed()) {
+                            await metaMaskPage.waitForTimeout(2000);
+                        }
+                    } catch (e) {
+                        console.log("Page closed during timeout after transfer request dialog");
+                        return;
+                    }
+                    currentAttempt++;
+                    continue;
+                }
+
+                // Check for final confirm button
+                const confirmButton = metaMaskPage.locator('[data-testid="confirm-footer-button"]');
+                const isConfirmVisible = await confirmButton.isVisible().catch(() => false);
+                if (isConfirmVisible) {
+                    console.log(`Attempt ${currentAttempt + 1}: Final confirm button found`);
+                    await confirmButton.click().catch(e => console.log("Failed to click confirm button:", e));
+                    console.log("Transaction confirmed successfully");
+                    break;
+                }
+
+                // Check for rejection button (in case of insufficient funds)
+                const rejectButton = metaMaskPage.locator('[data-testid="reject-footer-button"]');
+                const isRejectVisible = await rejectButton.isVisible().catch(() => false);
+                if (isRejectVisible) {
+                    console.log(`Attempt ${currentAttempt + 1}: Rejection button found (possibly insufficient funds)`);
+                    await rejectButton.click().catch(e => console.log("Failed to click reject button:", e));
+                    console.log("Transaction rejected");
+                    break;
+                }
+
+                // If none of the expected elements are found, wait and try again
+                console.log(`Attempt ${currentAttempt + 1}: No expected elements found, waiting...`);
+
+                // Check if page is closed before attempting to wait
+                if (metaMaskPage.isClosed()) {
+                    console.log("MetaMask page closed during wait - skipping timeout");
+                    return;
+                }
+
+                // Use a safer approach to wait that won't throw if the page closes
+                try {
+                    // Use a shorter timeout to reduce waiting time if page is closed
+                    await Promise.race([
+                        metaMaskPage.waitForTimeout(2000).catch(() => {
+                        }),
+                        new Promise(resolve => setTimeout(resolve, 2000))
+                    ]);
+                } catch (e) {
+                    console.log("Timeout handled gracefully");
+                    // Don't return, just continue with the next attempt
+                }
+                currentAttempt++;
+
+
+            } catch (error) {
+                console.log(`Attempt ${currentAttempt + 1} failed:`, error);
+                currentAttempt++;
+
+                // Check if page is closed before attempting to wait
+                if (metaMaskPage.isClosed()) {
+                    console.log("MetaMask page closed after error - skipping timeout");
+                    return;
+                }
+
+                // Use a safer approach to wait that won't throw if the page closes
+                try {
+                    // Use Promise.race to handle potential page closure during timeout
+                    await Promise.race([
+                        metaMaskPage.waitForTimeout(1000).catch(() => {
+                        }),
+                        new Promise(resolve => setTimeout(resolve, 1000))
+                    ]);
+                } catch (e) {
+                    console.log("Error timeout handled gracefully");
+                    // Continue with next attempt
+                }
             }
-          }
-
-          // Wait for the next dialog to appear
-          try {
-            if (!metaMaskPage.isClosed()) {
-              await metaMaskPage.waitForTimeout(2000);
-            }
-          } catch (e) {
-            console.log("Page closed during timeout after transfer request dialog");
-            return;
-          }
-          currentAttempt++;
-          continue;
         }
 
-        // Check for final confirm button
-        const confirmButton = metaMaskPage.locator('[data-testid="confirm-footer-button"]');
-        const isConfirmVisible = await confirmButton.isVisible().catch(() => false);
-        if (isConfirmVisible) {
-          console.log(`Attempt ${currentAttempt + 1}: Final confirm button found`);
-          await confirmButton.click().catch(e => console.log("Failed to click confirm button:", e));
-          console.log("Transaction confirmed successfully");
-          break;
+        // Wait for the MetaMask page to close or navigate away
+        if (!metaMaskPage.isClosed()) {
+            await metaMaskPage.waitForEvent('close', {timeout: 5000}).catch(() => {
+                console.log("MetaMask page didn't close as expected, continuing...");
+            });
+        } else {
+            console.log("MetaMask page already closed");
         }
 
-        // Check for rejection button (in case of insufficient funds)
-        const rejectButton = metaMaskPage.locator('[data-testid="reject-footer-button"]');
-        const isRejectVisible = await rejectButton.isVisible().catch(() => false);
-        if (isRejectVisible) {
-          console.log(`Attempt ${currentAttempt + 1}: Rejection button found (possibly insufficient funds)`);
-          await rejectButton.click().catch(e => console.log("Failed to click reject button:", e));
-          console.log("Transaction rejected");
-          break;
-        }
-
-        // If none of the expected elements are found, wait and try again
-        console.log(`Attempt ${currentAttempt + 1}: No expected elements found, waiting...`);
-
-        // Check if page is closed before attempting to wait
-        if (metaMaskPage.isClosed()) {
-          console.log("MetaMask page closed during wait - skipping timeout");
-          return;
-        }
-
-        // Use a safer approach to wait that won't throw if the page closes
-        try {
-          // Use a shorter timeout to reduce waiting time if page is closed
-          await Promise.race([
-            metaMaskPage.waitForTimeout(2000).catch(() => { }),
-            new Promise(resolve => setTimeout(resolve, 2000))
-          ]);
-        } catch (e) {
-          console.log("Timeout handled gracefully");
-          // Don't return, just continue with the next attempt
-        }
-        currentAttempt++;
-
-
-
-      } catch (error) {
-        console.log(`Attempt ${currentAttempt + 1} failed:`, error);
-        currentAttempt++;
-
-        // Check if page is closed before attempting to wait
-        if (metaMaskPage.isClosed()) {
-          console.log("MetaMask page closed after error - skipping timeout");
-          return;
-        }
-
-        // Use a safer approach to wait that won't throw if the page closes
-        try {
-          // Use Promise.race to handle potential page closure during timeout
-          await Promise.race([
-            metaMaskPage.waitForTimeout(1000).catch(() => { }),
-            new Promise(resolve => setTimeout(resolve, 1000))
-          ]);
-        } catch (e) {
-          console.log("Error timeout handled gracefully");
-          // Continue with next attempt
-        }
-      }
+    } catch (error) {
+        console.error("Error in handleMetaMaskNetworkAndConfirm:", error);
+        // Don't throw the error, just log it and continue
+        // This prevents the test from failing if MetaMask interaction fails
     }
-
-    // Wait for the MetaMask page to close or navigate away
-    if (!metaMaskPage.isClosed()) {
-      await metaMaskPage.waitForEvent('close', { timeout: 5000 }).catch(() => {
-        console.log("MetaMask page didn't close as expected, continuing...");
-      });
-    } else {
-      console.log("MetaMask page already closed");
-    }
-
-  } catch (error) {
-    console.error("Error in handleMetaMaskNetworkAndConfirm:", error);
-    // Don't throw the error, just log it and continue
-    // This prevents the test from failing if MetaMask interaction fails
-  }
 }
 
 // Helper function to upload a file
 export async function uploadFile(page: Page, filePath: string, dropzoneSelector: string = '[data-testid="file-upload-dropzone"]'): Promise<void> {
-  console.log("Waiting for file upload dropzone to be visible...");
-  await page.waitForSelector(dropzoneSelector, { state: 'visible', timeout: 10000 });
-  console.log("File upload dropzone is visible");
+    console.log("Waiting for file upload dropzone to be visible...");
+    await page.waitForSelector(dropzoneSelector, {state: 'visible', timeout: 10000});
+    console.log("File upload dropzone is visible");
 
-  const fileChooserPromise = page.waitForEvent('filechooser');
-  await page.click(dropzoneSelector);
-  const fileChooser = await fileChooserPromise;
-  await fileChooser.setFiles(filePath);
+    const fileChooserPromise = page.waitForEvent('filechooser');
+    await page.click(dropzoneSelector);
+    const fileChooser = await fileChooserPromise;
+    await fileChooser.setFiles(filePath);
 
-  console.log("File uploaded successfully");
+    console.log("File uploaded successfully");
 }
 
 // Helper function to wait for MetaMask popup and return the page
 export async function waitForMetaMaskPopup(context: BrowserContext): Promise<Page> {
-  const metaMaskPromise = context.waitForEvent("page");
-  await metaMaskPromise;
-  return context.pages()[1];
+    const metaMaskPromise = context.waitForEvent("page");
+    await metaMaskPromise;
+    return context.pages()[1];
 }
 
 // Helper function to close upload dialog
 export async function closeUploadDialog(page: Page): Promise<void> {
-  console.log("Waiting for clear completed button to appear...");
-  await page.waitForSelector('[data-testid="clear-completed-button"]', { state: 'visible', timeout: 10000 });
-  await page.click('[data-testid="clear-completed-button"]');
+    console.log("Waiting for clear completed button to appear...");
+    await page.waitForSelector('[data-testid="clear-completed-button"]', {state: 'visible', timeout: 10000});
+    await page.click('[data-testid="clear-completed-button"]');
 
-  console.log("Waiting for close upload dialog to appear");
-  await page.waitForSelector('[data-testid="close-upload-dialog-button"]', { state: 'visible', timeout: 10000 });
-  await page.click('[data-testid="close-upload-dialog-button"]');
+    console.log("Waiting for close upload dialog to appear");
+    await page.waitForSelector('[data-testid="close-upload-dialog-button"]', {state: 'visible', timeout: 10000});
+    await page.click('[data-testid="close-upload-dialog-button"]');
 
-  console.log("Clicked close upload dialog button");
+    console.log("Clicked close upload dialog button");
 }
 
 
 // Helper function to sign a document
 export async function witnessDocument(page: Page, context: BrowserContext): Promise<void> {
-  // Wait longer for the UI to stabilize
-  await page.waitForTimeout(12000);
-
-  try {
-    console.log("Waiting for witness button to appear...");
-    await page.waitForSelector('[data-testid="witness-action-button"]', { state: 'visible', timeout: 15000 });
-    console.log("Witness button is visible");
-
-    // Check if MetaMask page already exists
-    const existingMetaMaskPage = context.pages().find(page => page.url().includes('extension'));
+    // Wait longer for the UI to stabilize
+    await page.waitForTimeout(12000);
 
     try {
-      if (existingMetaMaskPage && !existingMetaMaskPage.isClosed()) {
-        console.log("MetaMask page already exists, using existing page");
-        await page.click('[data-testid="witness-action-button"]').catch(e => {
-          console.log("Error clicking witness button, will try to continue:", e);
-        });
+        console.log("Waiting for witness button to appear...");
+        await page.waitForSelector('[data-testid="witness-action-button"]', {state: 'visible', timeout: 15000});
+        console.log("Witness button is visible");
 
-        // Use a non-throwing version of handleMetaMaskNetworkAndConfirm
-        await handleMetaMaskNetworkAndConfirm(context, true, existingMetaMaskPage).catch(e => {
-          console.log("MetaMask interaction failed but continuing with test:", e);
-        });
-      } else {
-        console.log("Waiting for new MetaMask page...");
+        // Check if MetaMask page already exists
+        const existingMetaMaskPage = context.pages().find(page => page.url().includes('extension'));
+
         try {
-          // Click the button first
-          await page.click('[data-testid="witness-action-button"]');
-          console.log("Clicked witness button, waiting for MetaMask popup...");
+            if (existingMetaMaskPage && !existingMetaMaskPage.isClosed()) {
+                console.log("MetaMask page already exists, using existing page");
+                await page.click('[data-testid="witness-action-button"]').catch(e => {
+                    console.log("Error clicking witness button, will try to continue:", e);
+                });
 
-          // Then wait for the popup with a timeout
-          const metaMaskPromise = Promise.race([
-            context.waitForEvent("page", { timeout: 10000 }),
-            new Promise((_, reject) => setTimeout(() => reject(new Error("Timeout waiting for MetaMask popup")), 10000))
-          ]);
+                // Use a non-throwing version of handleMetaMaskNetworkAndConfirm
+                await handleMetaMaskNetworkAndConfirm(context, true, existingMetaMaskPage).catch(e => {
+                    console.log("MetaMask interaction failed but continuing with test:", e);
+                });
+            } else {
+                console.log("Waiting for new MetaMask page...");
+                try {
+                    // Click the button first
+                    await page.click('[data-testid="witness-action-button"]');
+                    console.log("Clicked witness button, waiting for MetaMask popup...");
 
-          await metaMaskPromise.catch(e => {
-            console.log("Failed to get MetaMask popup, continuing anyway:", e);
-          });
+                    // Then wait for the popup with a timeout
+                    const metaMaskPromise = Promise.race([
+                        context.waitForEvent("page", {timeout: 10000}),
+                        new Promise((_, reject) => setTimeout(() => reject(new Error("Timeout waiting for MetaMask popup")), 10000))
+                    ]);
 
-          console.log("MetaMask popup received,,,");
-          // Try to handle MetaMask interaction but don't fail if it doesn't work
-          await handleMetaMaskNetworkAndConfirm(context, true).catch(e => {
-            console.log("MetaMask interaction failed but continuing with test:", e);
-          });
-        } catch (clickError) {
-          console.log("Error during witness button click or MetaMask popup wait:", clickError);
-          // Continue despite errors
+                    await metaMaskPromise.catch(e => {
+                        console.log("Failed to get MetaMask popup, continuing anyway:", e);
+                    });
+
+                    console.log("MetaMask popup received,,,");
+                    // Try to handle MetaMask interaction but don't fail if it doesn't work
+                    await handleMetaMaskNetworkAndConfirm(context, true).catch(e => {
+                        console.log("MetaMask interaction failed but continuing with test:", e);
+                    });
+                } catch (clickError) {
+                    console.log("Error during witness button click or MetaMask popup wait:", clickError);
+                    // Continue despite errors
+                }
+            }
+
+            // Download the witnessed document to verify it was witnessed properly
+            try {
+                console.log("Attempting to download the witnessed document for verification...");
+
+                // Wait for the witness process to complete (give it some time)
+                await page.waitForTimeout(3000);
+
+                // Check if the download button is available
+                const downloadButton = page.locator('[data-testid="download-aqua-tree-button"]');
+                const isDownloadButtonVisible = await downloadButton.isVisible().catch(() => false);
+
+                if (isDownloadButtonVisible) {
+                    console.log("Download button found - document was successfully witnessed");
+                    await downloadButton.click().catch(e => {
+                        console.log("Error clicking download button, but document appears to be witnessed:", e);
+                    });
+                    console.log("Witnessed document downloaded for verification");
+                } else {
+                    console.log("Download button not found - document may not have been witnessed successfully");
+                    // Take a screenshot for debugging
+                    await page.screenshot({path: 'witness-verification-failed.png'}).catch(() => {
+                    });
+                }
+            } catch (verificationError) {
+                console.log("Error during witness verification:", verificationError);
+                // Continue despite verification errors
+            }
+        } catch (metaMaskError) {
+            console.log("Error during MetaMask interaction, continuing with test:", metaMaskError);
+            // Continue with the test despite MetaMask errors
         }
-      }
-
-      // Download the witnessed document to verify it was witnessed properly
-      try {
-        console.log("Attempting to download the witnessed document for verification...");
-
-        // Wait for the witness process to complete (give it some time)
-        await page.waitForTimeout(3000);
-
-        // Check if the download button is available
-        const downloadButton = page.locator('[data-testid="download-aqua-tree-button"]');
-        const isDownloadButtonVisible = await downloadButton.isVisible().catch(() => false);
-
-        if (isDownloadButtonVisible) {
-          console.log("Download button found - document was successfully witnessed");
-          await downloadButton.click().catch(e => {
-            console.log("Error clicking download button, but document appears to be witnessed:", e);
-          });
-          console.log("Witnessed document downloaded for verification");
-        } else {
-          console.log("Download button not found - document may not have been witnessed successfully");
-          // Take a screenshot for debugging
-          await page.screenshot({ path: 'witness-verification-failed.png' }).catch(() => { });
-        }
-      } catch (verificationError) {
-        console.log("Error during witness verification:", verificationError);
-        // Continue despite verification errors
-      }
-    } catch (metaMaskError) {
-      console.log("Error during MetaMask interaction, continuing with test:", metaMaskError);
-      // Continue with the test despite MetaMask errors
+    } catch (error) {
+        console.log("Error during witness process, likely insufficient funds or UI issues. Skipping:", error);
+        console.log("Continuing with test despite witness failure");
+        // Continue with the test despite any errors
     }
-  } catch (error) {
-    console.log("Error during witness process, likely insufficient funds or UI issues. Skipping:", error);
-    console.log("Continuing with test despite witness failure");
-    // Continue with the test despite any errors
-  }
 
-  console.log("Witness document step completed (with or without success)");
+    console.log("Witness document step completed (with or without success)");
 }
 
 // Helper function to sign a document
 export async function signDocument(page: Page, context: BrowserContext): Promise<void> {
-  console.log("Waiting for sign button to appear...");
-  await page.waitForSelector('[data-testid="sign-action-button-0"]', { state: 'visible' });
-  console.log("Sign button is visible");
+    console.log("Waiting for sign button to appear...");
+    await page.waitForSelector('[data-testid="sign-action-button-0"]', {state: 'visible'});
+    console.log("Sign button is visible");
 
-  const metaMaskPromise = context.waitForEvent("page");
-  await page.click('[data-testid="sign-action-button-0"]');
-  console.log("Clicked sign button, waiting for MetaMask popup...");
+    const metaMaskPromise = context.waitForEvent("page");
+    await page.click('[data-testid="sign-action-button-0"]');
+    console.log("Clicked sign button, waiting for MetaMask popup...");
 
-  await metaMaskPromise;
-  await handleMetaMaskNetworkAndConfirm(context, true);
+    await metaMaskPromise;
+    await handleMetaMaskNetworkAndConfirm(context, true);
 }
 
 // Helper function to download aqua tree
-export async function downloadAquaTree(page: Page, saveToDownloads : boolean ): Promise<void> {
-  // Create a downloads directory if it doesn't exist
-  const downloadsPath = path.join(__dirname, 'downloads');
-  if (!fs.existsSync(downloadsPath)) {
-    fs.mkdirSync(downloadsPath);
-  }
+export async function downloadAquaTree(page: Page, saveToDownloads: boolean): Promise<void> {
+    // Create a downloads directory if it doesn't exist
+    const downloadsPath = path.join(__dirname, 'downloads');
+    if (!fs.existsSync(downloadsPath)) {
+        fs.mkdirSync(downloadsPath);
+    }
 
-  // Set up download listener before clicking the button
-  const downloadPromise = page.waitForEvent('download');
+    // Set up download listener before clicking the button
+    const downloadPromise = page.waitForEvent('download');
 
-  // Click the download button
-  //download-aqua-tree-button-1
-  await page.waitForSelector('[data-testid="download-aqua-tree-button-0"]', { state: 'visible' });
-  await page.click('[data-testid="download-aqua-tree-button-0"]');
+    // Click the download button
+    //download-aqua-tree-button-1
+    await page.waitForSelector('[data-testid="download-aqua-tree-button-0"]', {state: 'visible'});
+    await page.click('[data-testid="download-aqua-tree-button-0"]');
 
-  // Wait for the download to start
-  const download = await downloadPromise;
-  console.log(`Download started: ${download.suggestedFilename()}`);
+    // Wait for the download to start
+    const download = await downloadPromise;
+    console.log(`Download started: ${download.suggestedFilename()}`);
 
-  // Save the downloaded file to the specified path
-  const filePath = path.join(downloadsPath, download.suggestedFilename());
-  await download.saveAs(filePath);
+    // Save the downloaded file to the specified path
+    const filePath = path.join(downloadsPath, download.suggestedFilename());
+    await download.saveAs(filePath);
 
-  console.log(`Download completed and saved to: ${filePath}`);
+    console.log(`Download completed and saved to: ${filePath}`);
 }
 
 // Helper function to create aqua sign form
 export async function createAquaSignForm(page: Page, context: BrowserContext, filePath: string, signerAddress?: string): Promise<void> {
-  await page.click('[data-testid="create-document-signature"]');
-  console.log("clicked aqua sign");
+    await page.click('[data-testid="create-document-signature"]');
+    console.log("clicked aqua sign");
 
-  // Upload document
-  await uploadFile(page, filePath, '[data-testid="input-document"]');
+    // Upload document
+    await uploadFile(page, filePath, '[data-testid="input-document"]');
 
-  // Configure signers
-  await page.click('[data-testid="multiple_values_signers"]');
-  console.log("clicked multiple values signers");
-
-  const metaMaskAdr: string = await page.locator('[data-testid="input-sender"]').inputValue();
-  await page.fill('[data-testid="input-signers-0"]', metaMaskAdr);
-  console.log("filled aqua sign form");
-
-  if (signerAddress) {
+    // Configure signers
     await page.click('[data-testid="multiple_values_signers"]');
     console.log("clicked multiple values signers");
 
-    await page.fill('[data-testid="input-signers-1"]', signerAddress);
+    const metaMaskAdr: string = await page.locator('[data-testid="input-sender"]').inputValue();
+    await page.fill('[data-testid="input-signers-0"]', metaMaskAdr);
     console.log("filled aqua sign form");
-  }
 
-  // Submit form and handle MetaMask
-  const metamaskPromise = context.waitForEvent("page");
-  await page.click('[type="submit"]');
-  await metamaskPromise;
+    if (signerAddress) {
+        await page.click('[data-testid="multiple_values_signers"]');
+        console.log("clicked multiple values signers");
 
-  await handleMetaMaskNetworkAndConfirm(context, true);
+        await page.fill('[data-testid="input-signers-1"]', signerAddress);
+        console.log("filled aqua sign form");
+    }
+
+    // Submit form and handle MetaMask
+    const metamaskPromise = context.waitForEvent("page");
+    await page.click('[type="submit"]');
+    await metamaskPromise;
+
+    await handleMetaMaskNetworkAndConfirm(context, true);
 }
 
 // Helper function to handle signature creation and saving
 export async function importAquaChain(secondTestPage: Page, context: BrowserContext, shareUrl = ""): Promise<void> {
+    let url = ""
+    if (shareUrl.length == 0) {
+        url = `/app/shared-contracts`
+    } else {
+        url = shareUrl
+    }
 
-  let url = ""
-  if (shareUrl.length == 0) {
-    const baseUrl = process.env.BASE_URL || "http://localhost:5173";
-    url = `${baseUrl}/app/shared-contracts`
-  } else {
-    url = shareUrl
-  }
+    await secondTestPage.goto(url);
+    await secondTestPage.waitForLoadState('networkidle');
 
-  await secondTestPage.goto(url);
-  await secondTestPage.waitForLoadState('networkidle');
+    if (shareUrl.length == 0) {
+        await waitAndClick(secondTestPage, '[data-testid="open-shared-contract-button-0"]')
+    }
 
-  if (shareUrl.length == 0) {
-    await secondTestPage.waitForSelector('[data-testid="open-shared-contract-button-0"]', { state: 'visible', timeout: 10000 });
-    await secondTestPage.click('[data-testid="open-shared-contract-button-0"]')
-
-    await secondTestPage.waitForTimeout(2000);
-  }
-
-
-  await secondTestPage.waitForSelector('[data-testid="import-aqua-chain-1-button"]', { state: 'visible', timeout: 10000 });
-  await secondTestPage.click('[data-testid="import-aqua-chain-1-button"]')
-
-  await secondTestPage.waitForTimeout(2000);
-
+    await waitAndClick(secondTestPage, '[data-testid="import-aqua-chain-1-button"]');
 }
 
 export async function createAndSaveSignature(page: Page, context: BrowserContext): Promise<void> {
-  await page.getByText("Create Signature").waitFor({ state: 'visible' });
-  await page.getByText("Create Signature").click();
-  console.log("created signature");
+    await page.getByText("Create Signature").waitFor({state: 'visible'});
+    await page.getByText("Create Signature").click();
+    console.log("created signature");
 
-  await page.getByText("Save Signature").waitFor({ state: 'visible' });
-  await page.waitForSelector('[class="signature-canvas"]', { state: 'visible' });
-  await page.click('[class="signature-canvas"]');
+    await page.getByText("Save Signature").waitFor({state: 'visible'});
+    await page.waitForSelector('[class="signature-canvas"]', {state: 'visible'});
+    await page.click('[class="signature-canvas"]');
 
-  const metamaskPromise = context.waitForEvent("page");
-  await page.getByText("Save Signature").click();
-  await metamaskPromise;
+    const metamaskPromise = context.waitForEvent("page");
+    await page.getByText("Save Signature").click();
+    await metamaskPromise;
 
-  await handleMetaMaskNetworkAndConfirm(context, false);
-  console.log("signature saved");
+    await handleMetaMaskNetworkAndConfirm(context, false);
+    console.log("signature saved");
 }
 
 // Helper function to add signature to document
 export async function addSignatureToDocument(page: Page, context: BrowserContext): Promise<void> {
-  await page.getByText("Add Signature to document").waitFor({ state: 'visible' });
-  await page.getByText("Add Signature to document").click();
-  console.log("Add Signature to document");
+    await page.getByText("Add Signature to document").waitFor({state: 'visible'});
+    await page.getByText("Add Signature to document").click();
+    console.log("Add Signature to document");
 
-  await page.waitForSelector('[data-testid="pdf-canvas"]', { state: 'visible' });
-  await page.click('[data-testid="pdf-canvas"]');
-  console.log("Signature added to document");
+    await page.waitForSelector('[data-testid="pdf-canvas"]', {state: 'visible'});
+    await page.click('[data-testid="pdf-canvas"]');
+    console.log("Signature added to document");
 
-  const metamaskPromise = context.waitForEvent("page");
-  await page.getByText("Sign document").click();
-  console.log("Sign document button clicked");
+    const metamaskPromise = context.waitForEvent("page");
+    await page.getByText("Sign document").click();
+    console.log("Sign document button clicked");
 
-  await metamaskPromise;
-  await handleMetaMaskNetworkAndConfirm(context, false);
+    await metamaskPromise;
+    await handleMetaMaskNetworkAndConfirm(context, false);
 }
 
 /**
@@ -741,82 +733,77 @@ export async function addSignatureToDocument(page: Page, context: BrowserContext
  * Helper function to share a document with another user
  */
 export async function shareDocument(
-  page: Page,
-  context: BrowserContext,
-  recipientAddress: string
+    page: Page,
+    context: BrowserContext,
+    recipientAddress: string
 ): Promise<string> {
-  console.log(`Share document ....`)
-  let shareUrl = ""
+    console.log(`Share document ....`)
+    let shareUrl = ""
 
-  await page.waitForSelector('[data-testid="share-action-button-0"]', { state: 'visible', timeout: 10000 });
-  await page.click('[data-testid="share-action-button-0"]');
+    await waitAndClick(page, '[data-testid="share-action-button-0"]')
 
-  if (recipientAddress.length > 0) {
+    if (recipientAddress.length > 0) {
 
-    console.log(`Toggle specific wallet sharing`)
+        console.log(`Toggle specific wallet sharing`)
 
 
-    // Wait for the dialog to be fully loaded
-    await page.waitForSelector('text=Share with specific wallet', { state: 'visible', timeout: 5000 });
+        // Wait for the dialog to be fully loaded
+        await page.waitForSelector('text=Share with specific wallet', {state: 'visible', timeout: 5000});
 
-    // Toggle specific wallet sharing - try multiple selectors to find the switch
-    try {
-      // Method 1: Try to find the switch by its parent container
-      await page.locator('text=Share with specific wallet').locator('..').locator('button').click();
-    } catch (error) {
-      console.log('Method 1 failed, trying method 2...');
-      try {
-        // Method 2: Look for switch component directly
-        await page.locator('[role="switch"]').click();
-      } catch (error2) {
-        console.log('Method 2 failed, trying method 3...');
+        // Toggle specific wallet sharing - try multiple selectors to find the switch
         try {
-          // Method 3: Look for any button near the "Share with specific wallet" text
-          await page.locator('text=Share with specific wallet').locator('..//button').click();
-        } catch (error3) {
-          console.log('Method 3 failed, trying method 4...');
-          // Method 4: Use a more generic approach - find any clickable element in the switch container
-          const switchContainer = page.locator('text=Share with specific wallet').locator('..');
-          await switchContainer.locator('button, [role="switch"], [data-state]').first().click();
+            // Method 1: Try to find the switch by its parent container
+            await page.locator('text=Share with specific wallet').locator('..').locator('button').click();
+        } catch (error) {
+            console.log('Method 1 failed, trying method 2...');
+            try {
+                // Method 2: Look for switch component directly
+                await page.locator('[role="switch"]').click();
+            } catch (error2) {
+                console.log('Method 2 failed, trying method 3...');
+                try {
+                    // Method 3: Look for any button near the "Share with specific wallet" text
+                    await page.locator('text=Share with specific wallet').locator('..//button').click();
+                } catch (error3) {
+                    console.log('Method 3 failed, trying method 4...');
+                    // Method 4: Use a more generic approach - find any clickable element in the switch container
+                    const switchContainer = page.locator('text=Share with specific wallet').locator('..');
+                    await switchContainer.locator('button, [role="switch"], [data-state]').first().click();
+                }
+            }
         }
-      }
+
+        // Wait for the wallet address input to appear
+        await page.waitForSelector('input[placeholder="Enter wallet address"]', {state: 'visible', timeout: 5000});
+
+        // Enter recipient address
+        await page.locator('input[placeholder="Enter wallet address"]').fill(recipientAddress);
+    } else {
+        console.log(` sharing to everyone `)
     }
 
-    // Wait for the wallet address input to appear
-    await page.waitForSelector('input[placeholder="Enter wallet address"]', { state: 'visible', timeout: 5000 });
+    // Confirm sharing
+    // await page.getByTestId('share-modal-action-button-dialog').click();
+    await page.waitForSelector('[data-testid="share-modal-action-button-dialog"]', {state: 'visible', timeout: 10000});
+    await page.click('[data-testid="share-modal-action-button-dialog"]');
 
-    // Enter recipient address
-    await page.locator('input[placeholder="Enter wallet address"]').fill(recipientAddress);
-  } else {
-    console.log(` sharing to everyone `)
-  }
+    // Handle MetaMask confirmation if needed
+    // await handleMetaMaskNetworkAndConfirm(context);
+    console.log(` Shared Document Link to be visible `)
+    // Wait for success message - look for the shared document link section
+    await page.getByText('Shared Document Link').waitFor({state: 'visible', timeout: 30000});
 
-  // Confirm sharing
-  // await page.getByTestId('share-modal-action-button-dialog').click();
-  await page.waitForSelector('[data-testid="share-modal-action-button-dialog"]', { state: 'visible', timeout: 10000 });
-  await page.click('[data-testid="share-modal-action-button-dialog"]');
+    console.log(` copy share url logic  `)
+    // Method 1: Using getByTestId() and textContent() - Most common approach
+    shareUrl = await page.getByTestId('share-url').textContent() ?? "";
+    console.log('Share URL:', shareUrl);
 
-  // Handle MetaMask confirmation if needed
-  // await handleMetaMaskNetworkAndConfirm(context);
-  console.log(` Shared Document Link to be visible `)
-  // Wait for success message - look for the shared document link section
-  await page.getByText('Shared Document Link').waitFor({ state: 'visible', timeout: 30000 });
+    // Close the share dialog
+    // await page.getByTestId('share-cancel-action-button').click();
+    await page.waitForSelector('[data-testid="share-cancel-action-button"]', {state: 'visible', timeout: 10000});
+    await page.click('[data-testid="share-cancel-action-button"]');
 
-  console.log(` copy share url logic  `)
-  // Method 1: Using getByTestId() and textContent() - Most common approach
-  shareUrl = await page.getByTestId('share-url').textContent() ?? "";
-  console.log('Share URL:', shareUrl);
-
-
-  // Close the share dialog
-  // await page.getByTestId('share-cancel-action-button').click();
-  await page.waitForSelector('[data-testid="share-cancel-action-button"]', { state: 'visible', timeout: 10000 });
-  await page.click('[data-testid="share-cancel-action-button"]');
-
-
-
-
-  return shareUrl;
+    return shareUrl;
 }
 
 /**
@@ -838,9 +825,6 @@ export async function shareDocument(
 //   await page.waitForTimeout(2000);
 
 
-
-
-
 // await page.getByTestId('open-shared-contract-button-0').waitFor({ state: 'visible' });
 // await page.getByTestId('open-shared-contract-button-0').click();
 // Look for any shared files
@@ -859,540 +843,522 @@ export async function shareDocument(
 
 
 export function generatePassword(length: number): string {
-  let result = '';
-  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-  const charactersLength = characters.length;
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
+    let result = '';
+    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    const charactersLength = characters.length;
+    for (let i = 0; i < length; i++) {
+        result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
+    return result;
 }
 
 // NEW FUNCTION ADDED - Helper function to switch to a test network
 async function switchToTestNetwork(metaMaskPage: any) {
-  try {
-    // Click on network dropdown
-    await metaMaskPage.click('[data-testid="network-display"]');
+    try {
+        // Click on network dropdown
+        await metaMaskPage.click('[data-testid="network-display"]');
 
-    // Wait for network list
-    await metaMaskPage.waitForSelector('[data-testid="network-list"]', { state: 'visible', timeout: 5000 });
+        // Wait for network list
+        await metaMaskPage.waitForSelector('[data-testid="network-list"]', {state: 'visible', timeout: 5000});
 
-    // Try to select Sepolia testnet or localhost
-    const networkOptions = [
-      '[data-testid="Sepolia-list-item"]',
-      '[data-testid="Localhost 8545-list-item"]',
-      'button:has-text("Sepolia")',
-      'button:has-text("Localhost")'
-    ];
+        // Try to select Sepolia testnet or localhost
+        const networkOptions = [
+            '[data-testid="Sepolia-list-item"]',
+            '[data-testid="Localhost 8545-list-item"]',
+            'button:has-text("Sepolia")',
+            'button:has-text("Localhost")'
+        ];
 
-    for (const networkSelector of networkOptions) {
-      try {
-        await metaMaskPage.waitForSelector(networkSelector, { state: 'visible', timeout: 2000 });
-        await metaMaskPage.click(networkSelector);
-        console.log(`Switched to test network: ${networkSelector}`);
-        return;
-      } catch {
-        continue;
-      }
+        for (const networkSelector of networkOptions) {
+            try {
+                await metaMaskPage.waitForSelector(networkSelector, {state: 'visible', timeout: 2000});
+                await metaMaskPage.click(networkSelector);
+                console.log(`Switched to test network: ${networkSelector}`);
+                return;
+            } catch {
+                continue;
+            }
+        }
+
+        // If no test networks found, close the dropdown
+        await metaMaskPage.press('[data-testid="network-list"]', 'Escape');
+
+    } catch (error) {
+        console.log("Could not switch network:", error);
     }
-
-    // If no test networks found, close the dropdown
-    await metaMaskPage.press('[data-testid="network-list"]', 'Escape');
-
-  } catch (error) {
-    console.log("Could not switch network:", error);
-  }
 }
 
 
 export async function fundWallet(walletToFund: string) {
 
-  try {
+    try {
 
-    const funderPrivateKey = process.env.PREFUNDED_WALLET_PRIVATEKEY || "" //'0x...'; // Store securely!
-    const alchemyKey = process.env.ALCHEMY_PROJECT_ID
+        const funderPrivateKey = process.env.PREFUNDED_WALLET_PRIVATEKEY || "" //'0x...'; // Store securely!
+        const alchemyKey = process.env.ALCHEMY_PROJECT_ID
 
-    // Debug environment variables
-    console.log(`PREFUNDED_WALLET_PRIVATEKEY set: ${!!process.env.PREFUNDED_WALLET_PRIVATEKEY}`);
-    console.log(`ALCHEMY_PROJECT_ID set: ${!!process.env.ALCHEMY_PROJECT_ID}`);
+        // Debug environment variables
+        console.log(`PREFUNDED_WALLET_PRIVATEKEY set: ${!!process.env.PREFUNDED_WALLET_PRIVATEKEY}`);
+        console.log(`ALCHEMY_PROJECT_ID set: ${!!process.env.ALCHEMY_PROJECT_ID}`);
 
-    // Define network configuration explicitly
-    const sepoliaNetwork = {
-      name: 'sepolia',
-      chainId: 11155111
-    };
+        // Define network configuration explicitly
+        const sepoliaNetwork = {
+            name: 'sepolia',
+            chainId: 11155111
+        };
 
-    // Create provider with explicit network configuration
-    const alchemyURL = 'https://eth-sepolia.g.alchemy.com/v2/' + alchemyKey;
-    const provider = new ethers.JsonRpcProvider(alchemyURL, sepoliaNetwork);
+        // Create provider with explicit network configuration
+        const alchemyURL = 'https://eth-sepolia.g.alchemy.com/v2/' + alchemyKey;
+        const provider = new ethers.JsonRpcProvider(alchemyURL, sepoliaNetwork);
 
-    // Check if provider is connected and verify network
-    const network = await provider.getNetwork();
-    console.log(`Provider network: ${network.name}, chainId: ${network.chainId}`);
+        // Check if provider is connected and verify network
+        const network = await provider.getNetwork();
+        console.log(`Provider network: ${network.name}, chainId: ${network.chainId}`);
 
-    // Verify we're on Sepolia (using Number() to avoid BigInt issues)
-    const sepoliaChainId = 11155111;
-    const currentChainId = Number(network.chainId);
+        // Verify we're on Sepolia (using Number() to avoid BigInt issues)
+        const sepoliaChainId = 11155111;
+        const currentChainId = Number(network.chainId);
 
-    if (currentChainId !== sepoliaChainId) {
-      console.error(`Wrong network detected: ${currentChainId}. Expected Sepolia (${sepoliaChainId})`);
-      throw new Error(`Wrong network detected: ${currentChainId}. Expected Sepolia (${sepoliaChainId})`);
+        if (currentChainId !== sepoliaChainId) {
+            console.error(`Wrong network detected: ${currentChainId}. Expected Sepolia (${sepoliaChainId})`);
+            throw new Error(`Wrong network detected: ${currentChainId}. Expected Sepolia (${sepoliaChainId})`);
+        }
+
+        const blockNumber = await provider.getBlockNumber();
+        console.log(`Provider connected: true, latest block: ${blockNumber}`);
+
+        console.log(`funderPrivateKey ${funderPrivateKey ? '(set)' : '(empty)'} -- alchemyURL ${alchemyURL}`)
+
+        // Check the balance of the funder wallet
+        const funderWallet = new ethers.Wallet(funderPrivateKey, provider);
+        const funderBalance = await provider.getBalance(funderWallet.address);
+        console.log(`Funder wallet address: ${funderWallet.address} -- Balance: ${ethers.formatEther(funderBalance)} ETH`);
+
+        // Ensure the wallet has enough funds
+        if (funderBalance < ethers.parseEther('0.0006')) {
+            console.error('Insufficient funds in funder wallet');
+            throw new Error(`Insufficient funds in funder wallet: ${ethers.formatEther(funderBalance)} ETH`);
+        }
+
+        // Send ETH
+        const tx = await funderWallet.sendTransaction({
+            to: walletToFund,
+            value: ethers.parseEther('0.0005'),
+        });
+
+        await tx.wait();
+        console.log(`Funded ${walletToFund} in tx ${tx.hash}`);
+    } catch (e) {
+        console.log(`fundWallet Error ${e}`)
     }
-
-    const blockNumber = await provider.getBlockNumber();
-    console.log(`Provider connected: true, latest block: ${blockNumber}`);
-
-    console.log(`funderPrivateKey ${funderPrivateKey ? '(set)' : '(empty)'} -- alchemyURL ${alchemyURL}`)
-
-    // Check the balance of the funder wallet
-    const funderWallet = new ethers.Wallet(funderPrivateKey, provider);
-    const funderBalance = await provider.getBalance(funderWallet.address);
-    console.log(`Funder wallet address: ${funderWallet.address} -- Balance: ${ethers.formatEther(funderBalance)} ETH`);
-
-    // Ensure the wallet has enough funds
-    if (funderBalance < ethers.parseEther('0.0006')) {
-      console.error('Insufficient funds in funder wallet');
-      throw new Error(`Insufficient funds in funder wallet: ${ethers.formatEther(funderBalance)} ETH`);
-    }
-
-    // Send ETH
-    const tx = await funderWallet.sendTransaction({
-      to: walletToFund,
-      value: ethers.parseEther('0.0005'),
-    });
-
-    await tx.wait();
-    console.log(`Funded ${walletToFund} in tx ${tx.hash}`);
-  } catch (e) {
-    console.log(`fundWallet Error ${e}`)
-  }
 
 }
 
 export async function registerNewMetaMaskWallet(): Promise<RegisterMetaMaskResponse> {
-  const metamaskPath = path.join(__dirname, 'metamask-extension');
-  console.log(`metamaskPath: ${metamaskPath}`)
+    const metamaskPath = path.join(__dirname, 'metamask-extension');
+    console.log(`metamaskPath: ${metamaskPath}`)
 
-  // const isCI = process.env.CI === 'true';
-  const userDataDir = '';
-  const context = await chromium.launchPersistentContext(userDataDir, {
-    headless: false,//isCI,
-    channel: 'chromium',
-    args: [
-      `--disable-extensions-except=${metamaskPath}`,
-      `--load-extension=${metamaskPath}`,
-    ],
-    // Increase timeouts for MetaMask operations
-    timeout: 60000
-  });
+    // const isCI = process.env.CI === 'true';
+    const userDataDir = '';
+    const context = await chromium.launchPersistentContext(userDataDir, {
+        headless: false,//isCI,
+        channel: 'chromium',
+        args: [
+            `--disable-extensions-except=${metamaskPath}`,
+            `--load-extension=${metamaskPath}`,
+        ],
+        // Increase timeouts for MetaMask operations
+        timeout: 60000
+    });
 
-  try {
-    console.log(`context: ${JSON.stringify(context, null, 4)}`)
-
-
-    // Wait for MetaMask page to open
-    await context.waitForEvent("page", { timeout: 30000 })
-    console.log(`context.waitForEvent("page")`)
-
-    // Get the MetaMask page - it should be the second page opened
-    const pages = context.pages();
-    if (pages.length < 2) {
-      throw new Error("MetaMask page not found");
-    }
-
-    const metaMaskPage = pages[1];
-    console.log(`metaMaskPage: ${JSON.stringify(metaMaskPage, null, 4)}`)
-
-    // Wait for page to load
-    await metaMaskPage.waitForLoadState("load", { timeout: 30000 });
-
-    // Setup page - accept terms and create wallet
-    console.log("Setting up MetaMask - accepting terms")
-    await metaMaskPage.waitForSelector('[data-testid="onboarding-terms-checkbox"]', { timeout: 30000 });
-    await metaMaskPage.click('[data-testid="onboarding-terms-checkbox"]')
-    await metaMaskPage.click('[data-testid="onboarding-create-wallet"]')
-
-    // Decline telemetry data collection
-    console.log("Declining telemetry data collection")
-    await metaMaskPage.waitForSelector('[data-testid="metametrics-no-thanks"]', { state: 'visible', timeout: 30000 });
-    await metaMaskPage.click('[data-testid="metametrics-no-thanks"]')
-
-    // Set up password
-    console.log("Setting up password")
-    let myNewPassword = generatePassword(15)
-    await metaMaskPage.waitForSelector('[data-testid="create-password-new"]', { state: 'visible', timeout: 30000 })
-    await metaMaskPage.fill('[data-testid="create-password-new"]', myNewPassword)
-    await metaMaskPage.fill('[data-testid="create-password-confirm"]', myNewPassword)
-    await metaMaskPage.click('[data-testid="create-password-terms"]')
-    await metaMaskPage.click('[data-testid="create-password-wallet"]')
-
-    // Skip wallet backup
-    console.log("Skipping wallet backup")
-    await metaMaskPage.waitForSelector('[data-testid="secure-wallet-later"]', { state: 'visible', timeout: 30000 });
-    await metaMaskPage.click('[data-testid="secure-wallet-later"]')
-    await metaMaskPage.waitForSelector('[data-testid="skip-srp-backup-popover-checkbox"]', { state: 'visible', timeout: 30000 });
-    await metaMaskPage.click('[data-testid="skip-srp-backup-popover-checkbox"]')
-    await metaMaskPage.click('[data-testid="skip-srp-backup"]')
-
-    // Complete onboarding
-    console.log("Completing onboarding")
-    await metaMaskPage.waitForSelector('[data-testid="onboarding-complete-done"]', { state: 'visible', timeout: 30000 });
-    await metaMaskPage.click('[data-testid="onboarding-complete-done"]')
-
-    // Complete tutorial
-    console.log("Completing tutorial")
-    await metaMaskPage.waitForSelector('[data-testid="pin-extension-next"]', { state: 'visible', timeout: 30000 });
-    await metaMaskPage.click('[data-testid="pin-extension-next"]')
-    await metaMaskPage.waitForSelector('[data-testid="pin-extension-done"]', { state: 'visible', timeout: 30000 });
-    await metaMaskPage.click('[data-testid="pin-extension-done"]')
-
-    // Wait for network display to be visible
-    console.log("Waiting for network display")
-    await metaMaskPage.waitForSelector('[data-testid="network-display"]', { state: 'visible', timeout: 30000 });
-
-    // Switch to a test network to avoid mainnet connection issues
-    console.log("Switching to test network")
-
-    
     try {
-      
-      // Try stop the not now popup
-      try {
-        await metaMaskPage.waitForSelector('[data-testid="not-now-button"]', { state: 'visible', timeout: 30000 });
-        await metaMaskPage.click('[data-testid="not-now-button"]')
-      } catch (error) {
-        console.log("No not now popup appeared or it was already dismissed");
-      }
-      
-      // Click on network selector
-      await metaMaskPage.click('[data-testid="network-display"]');
-      
-      // Pause execution for 30 seconds, use promise
-      // await new Promise(resolve => setTimeout(resolve, 30000));
+        console.log(`context: ${JSON.stringify(context, null, 4)}`)
 
-      // I want you to click the show test networks here
-      // This is how it looks like class="toggle-button toggle-button--off"
-      await metaMaskPage.click('[class="toggle-button toggle-button--off"]');
 
-      // Look for Sepolia test network and click it
-      // data-testid="Sepolia" click on this element
-      await metaMaskPage.click('[data-testid="Sepolia"]');
-      // const sepoliaSelector = 'button:has-text("Sepolia")';
-      // await metaMaskPage.waitForSelector(sepoliaSelector, { timeout: 10000 });
-      // await metaMaskPage.click(sepoliaSelector);
+        // Wait for MetaMask page to open
+        await context.waitForEvent("page", {timeout: 30000})
+        console.log(`context.waitForEvent("page")`)
 
-      console.log("Switched to Sepolia test network");
+        // Get the MetaMask page - it should be the second page opened
+        const pages = context.pages();
+        if (pages.length < 2) {
+            throw new Error("MetaMask page not found");
+        }
+
+        const metaMaskPage = pages[1];
+        console.log(`metaMaskPage: ${JSON.stringify(metaMaskPage, null, 4)}`)
+
+        // Wait for page to load
+        await metaMaskPage.waitForLoadState("load", {timeout: 30000});
+
+        // Setup page - accept terms and create wallet
+        console.log("Setting up MetaMask - accepting terms")
+        await metaMaskPage.waitForSelector('[data-testid="onboarding-terms-checkbox"]', {timeout: 30000});
+        await metaMaskPage.click('[data-testid="onboarding-terms-checkbox"]')
+        await metaMaskPage.click('[data-testid="onboarding-create-wallet"]')
+
+        // Decline telemetry data collection
+        console.log("Declining telemetry data collection")
+        await metaMaskPage.waitForSelector('[data-testid="metametrics-no-thanks"]', {state: 'visible', timeout: 30000});
+        await metaMaskPage.click('[data-testid="metametrics-no-thanks"]')
+
+        // Set up password
+        console.log("Setting up password")
+        let myNewPassword = generatePassword(15)
+        await metaMaskPage.waitForSelector('[data-testid="create-password-new"]', {state: 'visible', timeout: 30000})
+        await metaMaskPage.fill('[data-testid="create-password-new"]', myNewPassword)
+        await metaMaskPage.fill('[data-testid="create-password-confirm"]', myNewPassword)
+        await metaMaskPage.click('[data-testid="create-password-terms"]')
+        await metaMaskPage.click('[data-testid="create-password-wallet"]')
+
+        // Skip wallet backup
+        console.log("Skipping wallet backup")
+        await metaMaskPage.waitForSelector('[data-testid="secure-wallet-later"]', {state: 'visible', timeout: 30000});
+        await metaMaskPage.click('[data-testid="secure-wallet-later"]')
+        await metaMaskPage.waitForSelector('[data-testid="skip-srp-backup-popover-checkbox"]', {
+            state: 'visible',
+            timeout: 30000
+        });
+        await metaMaskPage.click('[data-testid="skip-srp-backup-popover-checkbox"]')
+        await metaMaskPage.click('[data-testid="skip-srp-backup"]')
+
+        // Complete onboarding
+        console.log("Completing onboarding")
+        await metaMaskPage.waitForSelector('[data-testid="onboarding-complete-done"]', {
+            state: 'visible',
+            timeout: 30000
+        });
+        await metaMaskPage.click('[data-testid="onboarding-complete-done"]')
+
+        // Complete tutorial
+        console.log("Completing tutorial")
+        await metaMaskPage.waitForSelector('[data-testid="pin-extension-next"]', {state: 'visible', timeout: 30000});
+        await metaMaskPage.click('[data-testid="pin-extension-next"]')
+        await metaMaskPage.waitForSelector('[data-testid="pin-extension-done"]', {state: 'visible', timeout: 30000});
+        await metaMaskPage.click('[data-testid="pin-extension-done"]')
+
+        // Wait for network display to be visible
+        console.log("Waiting for network display")
+        await metaMaskPage.waitForSelector('[data-testid="network-display"]', {state: 'visible', timeout: 30000});
+
+        // Switch to a test network to avoid mainnet connection issues
+        console.log("Switching to test network")
+
+
+        try {
+
+            // Try stop the not now popup
+            try {
+                await metaMaskPage.waitForSelector('[data-testid="not-now-button"]', {
+                    state: 'visible',
+                    timeout: 30000
+                });
+                await metaMaskPage.click('[data-testid="not-now-button"]')
+            } catch (error) {
+                console.log("No not now popup appeared or it was already dismissed");
+            }
+
+            // Click on network selector
+            await metaMaskPage.click('[data-testid="network-display"]');
+
+            // Pause execution for 30 seconds, use promise
+            // await new Promise(resolve => setTimeout(resolve, 30000));
+
+            // I want you to click the show test networks here
+            // This is how it looks like class="toggle-button toggle-button--off"
+            await metaMaskPage.click('[class="toggle-button toggle-button--off"]');
+
+            // Look for Sepolia test network and click it
+            // data-testid="Sepolia" click on this element
+            await metaMaskPage.click('[data-testid="Sepolia"]');
+            // const sepoliaSelector = 'button:has-text("Sepolia")';
+            // await metaMaskPage.waitForSelector(sepoliaSelector, { timeout: 10000 });
+            // await metaMaskPage.click(sepoliaSelector);
+
+            console.log("Switched to Sepolia test network");
+        } catch (error) {
+            console.log("Could not switch network, continuing with current network", error);
+        }
+
+        // Handle the popup that might appear asking to connect
+        console.log("Handling connection popup")
+        try {
+            await metaMaskPage.waitForSelector('[data-testid="not-now-button"]', {state: 'visible', timeout: 1000});
+            await metaMaskPage.click('[data-testid="not-now-button"]')
+        } catch (error) {
+            console.log("No connection popup appeared or it was already dismissed");
+        }
+
+        // Get wallet address
+        console.log("Getting wallet address")
+        await metaMaskPage.waitForSelector('[data-testid="account-menu-icon"]', {timeout: 30000})
+        await metaMaskPage.click('[data-testid="account-menu-icon"]')
+        await metaMaskPage.waitForSelector('[data-testid="account-list-item-menu-button"]', {timeout: 30000})
+        await metaMaskPage.click('[data-testid="account-list-item-menu-button"]')
+        await metaMaskPage.waitForSelector('[data-testid="account-list-menu-details"]', {timeout: 30000})
+        await metaMaskPage.click('[data-testid="account-list-menu-details"]')
+        await metaMaskPage.getByText("Details").waitFor({state: 'visible', timeout: 30000})
+        await metaMaskPage.getByText("Details").click()
+
+        // Get the wallet address
+        await metaMaskPage.waitForSelector('[class="mm-box mm-text qr-code__address-segments mm-text--body-md mm-box--margin-bottom-4 mm-box--color-text-default"]', {timeout: 30000})
+        const address = await metaMaskPage.locator('[class="mm-box mm-text qr-code__address-segments mm-text--body-md mm-box--margin-bottom-4 mm-box--color-text-default"]').textContent()
+
+        // Close the MetaMask page
+        await metaMaskPage.close()
+
+        console.log("Wallet finished!!! Adr: " + address)
+
+        if (address == null) {
+            throw Error(`Wallet address cannot be null`)
+        }
+        return new RegisterMetaMaskResponse(context, address);
     } catch (error) {
-      console.log("Could not switch network, continuing with current network", error);
+        console.error("Error in registerNewMetaMaskWallet:", error);
+        // Try to close the context to clean up resources
+        try {
+            await context.close();
+        } catch (closeError) {
+            console.error("Error closing context:", closeError);
+        }
+        throw error;
     }
-
-    // Handle the popup that might appear asking to connect
-    console.log("Handling connection popup")
-    try {
-      await metaMaskPage.waitForSelector('[data-testid="not-now-button"]', { state: 'visible', timeout: 10000 });
-      await metaMaskPage.click('[data-testid="not-now-button"]')
-    } catch (error) {
-      console.log("No connection popup appeared or it was already dismissed");
-    }
-
-    // Get wallet address
-    console.log("Getting wallet address")
-    await metaMaskPage.waitForSelector('[data-testid="account-menu-icon"]', { timeout: 30000 })
-    await metaMaskPage.click('[data-testid="account-menu-icon"]')
-    await metaMaskPage.waitForSelector('[data-testid="account-list-item-menu-button"]', { timeout: 30000 })
-    await metaMaskPage.click('[data-testid="account-list-item-menu-button"]')
-    await metaMaskPage.waitForSelector('[data-testid="account-list-menu-details"]', { timeout: 30000 })
-    await metaMaskPage.click('[data-testid="account-list-menu-details"]')
-    await metaMaskPage.getByText("Details").waitFor({ state: 'visible', timeout: 30000 })
-    await metaMaskPage.getByText("Details").click()
-
-    // Get the wallet address
-    await metaMaskPage.waitForSelector('[class="mm-box mm-text qr-code__address-segments mm-text--body-md mm-box--margin-bottom-4 mm-box--color-text-default"]', { timeout: 30000 })
-    const address = await metaMaskPage.locator('[class="mm-box mm-text qr-code__address-segments mm-text--body-md mm-box--margin-bottom-4 mm-box--color-text-default"]').textContent()
-
-    // Close the MetaMask page
-    await metaMaskPage.close()
-
-    console.log("Wallet finished!!! Adr: " + address)
-
-    if (address == null) {
-      throw Error(`Wallet address cannot be null`)
-    }
-    return new RegisterMetaMaskResponse(context, address);
-  } catch (error) {
-    console.error("Error in registerNewMetaMaskWallet:", error);
-    // Try to close the context to clean up resources
-    try {
-      await context.close();
-    } catch (closeError) {
-      console.error("Error closing context:", closeError);
-    }
-    throw error;
-  }
 }
 
-export async function registerNewMetaMaskWalletAndLogin(pathname: string = "/app"): Promise<RegisterMetaMaskResponse> {
-  const response = await registerNewMetaMaskWallet();
-  const context = response.context;
-  const testPage = context.pages()[0];
-  await testPage.waitForLoadState("load")
+export async function registerNewMetaMaskWalletAndLogin(url: string = "/app"): Promise<RegisterMetaMaskResponse> {
+    const response = await registerNewMetaMaskWallet();
+    const context = response.context;
+    const testPage = context.pages()[0];
+    await testPage.waitForLoadState("load")
 
-  // Get the BASE_URL from environment variables and navigate to it
-  const baseUrl = process.env.BASE_URL || "https://dev.inblock.io";
-  console.log(`BASE URL: ${baseUrl}`);
-  const url = `${baseUrl}/${pathname.startsWith('/') ? pathname.slice(1) : pathname}`;
-  console.log(`Navigating to: ${url}`);
-  await testPage.goto(url, { waitUntil: 'networkidle' })
+    await testPage.goto(url, {waitUntil: 'networkidle'})
 
-  console.log("Page loaded, looking for sign-in button...");
+    console.log("Page loaded, looking for sign-in button...");
 
-  try {
-    // Take a screenshot to help debug
-    // await testPage.screenshot({ path: 'page-before-login.png' });
-    // console.log("Screenshot saved as page-before-login.png");
+    try {
+        // Take a screenshot to help debug
+        // await testPage.screenshot({ path: 'page-before-login.png' });
+        // console.log("Screenshot saved as page-before-login.png");
 
-    // Look for any sign-in button with a more flexible approach
-    console.log("Looking for any sign-in button...");
+        // Look for any sign-in button with a more flexible approach
+        console.log("Looking for any sign-in button...");
 
-    // Wait for and click the sign-in button
-    //  await testPage.waitForSelector('[data-testid="sign-in-button-dialog"]', {state: 'visible', timeout: 60000})
-    //  console.log("Sign-in button found, clicking it...");
-    // Try different possible selectors for the sign-in button
-    const signInButtonSelectors = [
-      '[data-testid="sign-in-button-page"]',
-      'button:has-text("Sign In")',
-      'button:has-text("Connect")',
-      'button:has-text("Login")'
-    ];
+        // Wait for and click the sign-in button
+        //  await testPage.waitForSelector('[data-testid="sign-in-button-dialog"]', {state: 'visible', timeout: 60000})
+        //  console.log("Sign-in button found, clicking it...");
+        // Try different possible selectors for the sign-in button
+        const signInButtonSelectors = [
+            '[data-testid="sign-in-button-page"]',
+            'button:has-text("Sign In")',
+            'button:has-text("Connect")',
+            'button:has-text("Login")'
+        ];
 
-    // data-testid="unlock-submit" if this button is visible, fill in the password and click it, check the button first
-    // let isUnlockSubmitVisible = await testPage.waitForSelector('[data-testid="unlock-submit"]', { state: 'visible', timeout: 60000 })
-    // if (isUnlockSubmitVisible) {
-    //   await testPage.fill('[data-testid="password-input"]', process.env.METAMASK_PASSWORD || "")
-    //   await testPage.click('[data-testid="unlock-submit"]')
-    // }
+        // data-testid="unlock-submit" if this button is visible, fill in the password and click it, check the button first
+        // let isUnlockSubmitVisible = await testPage.waitForSelector('[data-testid="unlock-submit"]', { state: 'visible', timeout: 60000 })
+        // if (isUnlockSubmitVisible) {
+        //   await testPage.fill('[data-testid="password-input"]', process.env.METAMASK_PASSWORD || "")
+        //   await testPage.click('[data-testid="unlock-submit"]')
+        // }
 
-    let buttonFound = false;
+        let buttonFound = false;
 
-    // Click the sign-in button using the data-testid attribute
-    //  await testPage.click('[data-testid="sign-in-button-dialog"]');
-    // Try each selector until we find a visible button
-    for (const selector of signInButtonSelectors) {
-      try {
-        const isVisible = await testPage.isVisible(selector, { timeout: 5000 });
-        if (isVisible) {
-          console.log(`Found visible sign-in button with selector: ${selector}`);
-          await testPage.click(selector);
-          buttonFound = true;
-          break;
+        // Click the sign-in button using the data-testid attribute
+        //  await testPage.click('[data-testid="sign-in-button-dialog"]');
+        // Try each selector until we find a visible button
+        for (const selector of signInButtonSelectors) {
+            try {
+                const isVisible = await testPage.isVisible(selector, {timeout: 5000});
+                if (isVisible) {
+                    console.log(`Found visible sign-in button with selector: ${selector}`);
+                    await testPage.click(selector);
+                    buttonFound = true;
+                    break;
+                }
+            } catch (e) {
+                console.log(`Selector ${selector} not found or not visible`);
+            }
         }
-      } catch (e) {
-        console.log(`Selector ${selector} not found or not visible`);
-      }
+
+        if (!buttonFound) {
+            // If no button found with specific selectors, try to find any button that might be a sign-in button
+            console.log("No specific sign-in button found, looking for any button that might be for sign-in...");
+
+            // Take a screenshot to see what's on the page
+            // await testPage.screenshot({ path: 'page-no-button-found.png' });
+
+            // Force click the first button we find as a last resort
+            await testPage.click('button', {force: true});
+            console.log("Clicked first button found as fallback");
+        }
+        console.log("Clicked sign-in button, waiting for MetaMask popup...");
+    } catch (error) {
+        console.error("Error during login:", error);
+        await testPage.screenshot({path: 'login-error.png'});
+        throw error;
     }
 
-    if (!buttonFound) {
-      // If no button found with specific selectors, try to find any button that might be a sign-in button
-      console.log("No specific sign-in button found, looking for any button that might be for sign-in...");
 
-      // Take a screenshot to see what's on the page
-      // await testPage.screenshot({ path: 'page-no-button-found.png' });
+    console.log("Starting metamask...");
 
-      // Force click the first button we find as a last resort
-      await testPage.click('button', { force: true });
-      console.log("Clicked first button found as fallback");
+    // Check if MetaMask page already exists
+    let metamaskPage;
+    const pages = context.pages();
+    console.log(`Found ${pages.length} pages in context`);
+
+    for (const page of pages) {
+        console.log("Page URL:", page.url());
     }
-    console.log("Clicked sign-in button, waiting for MetaMask popup...");
-  } catch (error) {
-    console.error("Error during login:", error);
-    await testPage.screenshot({ path: 'login-error.png' });
-    throw error;
-  }
 
+    // Always wait for MetaMask popup - it should appear after clicking the sign-in button
+    try {
+        console.log("Waiting for MetaMask popup to appear...");
+        const metamaskPromise = context.waitForEvent("page", {timeout: 30000}); // Increased timeout
+        metamaskPage = await metamaskPromise;
+        console.log("MetaMask popup opened with URL:", metamaskPage.url());
+    } catch (error) {
+        console.error("Failed to detect MetaMask popup after 30 seconds:", error);
 
-  console.log("Starting metamask...");
-  
-  // Check if MetaMask page already exists
-  let metamaskPage;
-  const pages = context.pages();
-  console.log(`Found ${pages.length} pages in context`);
+        // Take screenshot of current state
+        await testPage.screenshot({path: 'metamask-popup-timeout.png'});
 
-  for (const page of pages) {
-    console.log("Page URL:", page.url());
-  }
+        // Check if MetaMask page appeared after timeout
+        const updatedPages = context.pages();
+        console.log(`After timeout: Found ${updatedPages.length} pages in context`);
 
-  // Always wait for MetaMask popup - it should appear after clicking the sign-in button
-  try {
-    console.log("Waiting for MetaMask popup to appear...");
-    const metamaskPromise = context.waitForEvent("page", { timeout: 30000 }); // Increased timeout
-    metamaskPage = await metamaskPromise;
-    console.log("MetaMask popup opened with URL:", metamaskPage.url());
-  } catch (error) {
-    console.error("Failed to detect MetaMask popup after 30 seconds:", error);
-    
-    // Take screenshot of current state
-    await testPage.screenshot({ path: 'metamask-popup-timeout.png' });
-    
-    // Check if MetaMask page appeared after timeout
-    const updatedPages = context.pages();
-    console.log(`After timeout: Found ${updatedPages.length} pages in context`);
-    
-    for (const page of updatedPages) {
-      console.log("After timeout - Page URL:", page.url());
+        for (const page of updatedPages) {
+            console.log("After timeout - Page URL:", page.url());
+        }
+
+        if (updatedPages.length > 1) {
+            console.log("MetaMask page found after timeout, continuing...");
+            metamaskPage = updatedPages[1];
+        } else {
+            throw new Error("MetaMask popup never appeared");
+        }
     }
-    
-    if (updatedPages.length > 1) {
-      console.log("MetaMask page found after timeout, continuing...");
-      metamaskPage = updatedPages[1];
-    } else {
-      throw new Error("MetaMask popup never appeared");
-    }
-  }
-  await metamaskPage.waitForSelector('[data-testid="confirm-btn"]', { state: 'visible' })
-  await metamaskPage.click('[data-testid="confirm-btn"]')
+    await metamaskPage.waitForSelector('[data-testid="confirm-btn"]', {state: 'visible'})
+    await metamaskPage.click('[data-testid="confirm-btn"]')
 
-  await metamaskPage.waitForSelector('[data-testid="confirm-footer-button"]', { state: 'visible' })
-  await metamaskPage.click('[data-testid="confirm-footer-button"]')
+    await metamaskPage.waitForSelector('[data-testid="confirm-footer-button"]', {state: 'visible'})
+    await metamaskPage.click('[data-testid="confirm-footer-button"]')
 
-  return response;
+    return response;
 }
 
 
 export async function findAndClickHighestSharedButton(page: Page): Promise<number | null> {
-  let highestCount = -1;
-  let currentCount = 0;
+    let highestCount = -1;
+    let currentCount = 0;
 
-  // Keep checking for higher numbered buttons until we don't find any
-  while (true) {
-    const selector = `[data-testid="shared-button-count-${currentCount}"]`;
+    // Keep checking for higher numbered buttons until we don't find any
+    while (true) {
+        const selector = `[data-testid="shared-button-count-${currentCount}"]`;
 
-    try {
-      // Check if the element exists (with a short timeout to avoid long waits)
-      await page.waitForSelector(selector, { state: 'attached', timeout: 1000 });
-      highestCount = currentCount;
-      currentCount++;
-    } catch (error) {
-      // Element doesn't exist, break the loop
-      break;
+        try {
+            // Check if the element exists (with a short timeout to avoid long waits)
+            await page.waitForSelector(selector, {state: 'attached', timeout: 1000});
+            highestCount = currentCount;
+            currentCount++;
+        } catch (error) {
+            // Element doesn't exist, break the loop
+            break;
+        }
     }
-  }
 
-  return highestCount
+    return highestCount
 
-  // If we found at least one button with count > 0, click the highest one
-  //   if (highestCount > 0) {
-  //     const highestSelector = `[data-testid="shared-button-count-${highestCount}"]`;
-  //     await page.waitForSelector(highestSelector, { state: 'visible', timeout: 10000 });
-  //     await page.click(highestSelector);
-  //     console.log(`Clicked button with highest count: ${highestCount}`);
-  //     return highestCount;
-  //   } else {
-  //     console.log('No shared-button-count elements with value > 0 found');
-  //     return null;
-  //   }
+    // If we found at least one button with count > 0, click the highest one
+    //   if (highestCount > 0) {
+    //     const highestSelector = `[data-testid="shared-button-count-${highestCount}"]`;
+    //     await page.waitForSelector(highestSelector, { state: 'visible', timeout: 10000 });
+    //     await page.click(highestSelector);
+    //     console.log(`Clicked button with highest count: ${highestCount}`);
+    //     return highestCount;
+    //   } else {
+    //     console.log('No shared-button-count elements with value > 0 found');
+    //     return null;
+    //   }
 }
-
 
 
 export async function createTemplate(page: Page): Promise<void> {
 // Try to find the button by data-testid first, then fallback to text
-  try {
-    await page.waitForSelector('[data-testid="action-create-template-button"]', { state: 'visible', timeout: 10000 });
-    await page.click('[data-testid="action-create-template-button"]');
-    console.log("Clicked create template button using data-testid");
-  } catch (error) {
-    console.log("Failed to find button by data-testid, trying by text...");
-    await page.waitForSelector('button:has-text("New Template")', { state: 'visible', timeout: 10000 });
-    await page.click('button:has-text("New Template")');
-    console.log("Clicked create template button using text selector");
-  }
-
-  console.log("Clicked create template button");
-  await page.fill('#title', 'Test Template');
-
-  // Add two fields
-  try {
-    await page.click('[data-testid="action-add-form-field-button"]');
-    console.log("Clicked add form field button using data-testid");
-  } catch (error) {
-    console.log("Failed to find add form field button by data-testid, trying by text...");
-    await page.click('button:has-text("Add Form Field")');
-    console.log("Clicked add form field button using text selector");
-  }
-  let fields = [
-    {
-      label: 'Name',
-      type: 'text',
-      required: true,
-      is_array: false
-    },
-    {
-      label: 'Age',
-      type: 'number',
-      required: true,
-      is_array: false
+    try {
+        await waitAndClick(page, '[data-testid="action-create-template-button"]')
+        console.log("Clicked create template button using data-testid");
+    } catch (error) {
+        console.log("Failed to find button by data-testid, trying by text...");
+        await waitAndClick(page, 'button:has-text("New Template"');
+        console.log("Clicked create template button using text selector");
     }
-  ]
 
-  console.log("Adding fields to template form");
-  // Fill the template form
-  try {
-    await page.fill(`[data-testid="field-label-0"]`, fields[0].label);
-    console.log("Filled first field label using data-testid");
-  } catch (error) {
-    console.log("Failed to find first field label by data-testid, trying by id...");
-    await page.fill(`#field-label-0`, fields[0].label);
-    console.log("Filled first field label using id selector");
-  }
-  // await page.selectOption(`[data-testid="field-type-0"]`, fields[0].type);
-  // await page.click(`[data-testid="field-required-0"]`);
+    console.log("Clicked create template button");
+    await page.fill('#title', 'Test Template');
 
-  console.log("First field added to template form");
-  try {
-    await page.click('[data-testid="action-add-form-field-button"]');
-    console.log("Clicked second add form field button using data-testid");
-  } catch (error) {
-    console.log("Failed to find second add form field button by data-testid, trying by text...");
-    await page.click('button:has-text("Add Form Field")');
-    console.log("Clicked second add form field button using text selector");
-  }
+    // Add two fields
+    await waitAndClick(page, '[data-testid="action-add-form-field-button"]');
+    console.log("Clicked add form field button using data-testid");
+    let fields = [
+        {
+            label: 'Name',
+            type: 'text',
+            required: true,
+            is_array: false
+        },
+        {
+            label: 'Age',
+            type: 'number',
+            required: true,
+            is_array: false
+        }
+    ]
 
-  // await page.waitForSelector('[data-testid="field-label-1"]', { state: 'visible', timeout: 10000 });
-  try {
-    await page.fill(`[data-testid="field-label-1"]`, fields[1].label);
-    console.log("Filled second field label using data-testid");
-  } catch (error) {
-    console.log("Failed to find second field label by data-testid, trying by id...");
-    await page.fill(`#field-label-1`, fields[1].label);
-    console.log("Filled second field label using id selector");
-  }
-  // await page.selectOption(`[data-testid="field-type-1"]`, fields[1].type);
-  // await page.click(`[data-testid="field-required-1"]`);
-  console.log("Second field added to template form");
+    console.log("Adding fields to template form");
+    // Fill the template form
+    try {
+        await page.fill(`[data-testid="field-label-0"]`, fields[0].label);
+        console.log("Filled first field label using data-testid");
+    } catch (error) {
+        console.log("Failed to find first field label by data-testid, trying by id...");
+        await page.fill(`#field-label-0`, fields[0].label);
+        console.log("Filled first field label using id selector");
+    }
+    // await page.selectOption(`[data-testid="field-type-0"]`, fields[0].type);
+    // await page.click(`[data-testid="field-required-0"]`);
 
-  // Save the form
-  try {
-    await page.click('[data-testid="save-form-action-button"]');
-    console.log("Clicked save form button using data-testid");
-  } catch (error) {
-    console.log("Failed to find save form button by data-testid, trying by text...");
-    await page.click('button:has-text("Save")');
-    console.log("Clicked save form button using text selector");
-  }
-  console.log("Template form saved");
+    console.log("First field added to template form");
+    await waitAndClick(page, '[data-testid="action-add-form-field-button"]')
 
-   
+    // await page.waitForSelector('[data-testid="field-label-1"]', { state: 'visible', timeout: 10000 });
+    try {
+        await page.fill(`[data-testid="field-label-1"]`, fields[1].label);
+        console.log("Filled second field label using data-testid");
+    } catch (error) {
+        console.log("Failed to find second field label by data-testid, trying by id...");
+        await page.fill(`#field-label-1`, fields[1].label);
+        console.log("Filled second field label using id selector");
+    }
+    // await page.selectOption(`[data-testid="field-type-1"]`, fields[1].type);
+    // await page.click(`[data-testid="field-required-1"]`);
+    console.log("Second field added to template form");
+
+    // Save the form
+    await waitAndClick(page, '[data-testid="save-form-action-button"]')
+    console.log("Template form saved");
 }
 
+export async function waitAndClick(page: Page, selector: string) {
+    await page.waitForSelector(selector, {state: 'visible'});
+    await page.click(selector);
+}
 
 class RegisterMetaMaskResponse {
 
+    constructor(context: BrowserContext, walletAddress: string) {
+        this.context = context;
+        this.walletAddress = walletAddress;
+    }
 
-  constructor(context: BrowserContext, walletAddress: string) {
-    this.context = context;
-    this.walletAddress = walletAddress;
-  }
-
-  context: BrowserContext;
-  walletAddress: string;
+    context: BrowserContext;
+    walletAddress: string;
 }
